requirements.txt
# requirements.txt

langchain-google-genai
langgraph
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic